+++ corerl/component/optimizers/LineSearchOpt/linesearchopt/__init__.py
@@ -1,0 +1,4 @@
+from corerl.component.optimizers.LineSearchOpt.linesearchopt import init as init
+from corerl.component.optimizers.LineSearchOpt.linesearchopt import search as search
+
+from .sls import Optimizer as Optimizer