+++ corerl/component/optimizers/factory.py
@@ -5,6 +5,7 @@ import torch
 from corerl.component.optimizers.torch_opts import (
     AdamConfig,
     ArmijoAdamConfig,
+    LSOConfig,
     OptimConfig,
     RmspropConfig,
     SgdConfig,
@@ -16,6 +17,7 @@ OptimizerConfig = (
     | AdamConfig
     | SgdConfig
     | ArmijoAdamConfig
+    | LSOConfig
 )
 
 def init_optimizer(