+++ test/small/data_pipeline/state_constructor/test_countdown.py
@@ -23,6 +23,7 @@ def test_int_countdown1():
         action_period=timedelta(minutes=4),
         obs_period=timedelta(minutes=1),
         kind='int',
+        normalize=False,
     )
 
     cd_adder = DecisionPointDetector(cd_cfg)
@@ -62,6 +63,34 @@ def test_int_countdown1():
     assert dfs_close(pf.data, expected)
     assert np.all(pf.decision_points == [0, 1, 0, 0, 0])
 
+def test_normalized_int_countdown1():
+    actions = pd.DataFrame({
+        'tag-1': [1, 1, 1, 1, 2, 2, 2],
+        'tag-2': [0, 0, 0, 0, 0, 0, 0],
+    })
+
+    pf = PipelineFrame(
+        data=actions,
+        data_mode=DataMode.ONLINE,
+    )
+
+    cd_cfg = CountdownConfig(
+        action_period=timedelta(minutes=4),
+        obs_period=timedelta(minutes=1),
+        kind='int',
+        normalize=True,
+    )
+
+    cd_adder = DecisionPointDetector(cd_cfg)
+    pf = cd_adder(pf)
+
+    expected = pd.DataFrame({
+        'tag-1':         [1, 1, 1, 1, 2, 2, 2],
+        'tag-2':         [0, 0, 0, 0, 0, 0, 0],
+        'countdown.[0]': [1., 0.75, 0.5, 0.25, 1., 0.75, 0.5],
+    })
+    assert dfs_close(pf.data, expected)
+    assert np.all(pf.decision_points == [1, 0, 0, 0, 1, 0, 0])
 
 def test_int_countdown2():
     """
@@ -83,6 +112,7 @@ def test_int_countdown2():
         action_period=timedelta(minutes=4),
         obs_period=timedelta(minutes=1),
         kind='int',
+        normalize=False,
     )
 
     cd_adder = DecisionPointDetector(cd_cfg)
@@ -116,6 +146,7 @@ def test_int_countdown3():
         action_period=timedelta(minutes=4),
         obs_period=timedelta(minutes=1),
         kind='int',
+        normalize=False,
     )
 
     cd_adder = DecisionPointDetector(cd_cfg)
@@ -143,6 +174,7 @@ def test_two_clock1():
         action_period=timedelta(minutes=4),
         obs_period=timedelta(minutes=1),
         kind='two_clock',
+        normalize=False,
     )
 
     cd_adder = DecisionPointDetector(cd_cfg)